+++ test/translator/translator.translate.returnObject.spec.js
@@ -3,6 +3,7 @@ import ResourceStore from '../../src/ResourceStore.js';
 import LanguageUtils from '../../src/LanguageUtils';
 import PluralResolver from '../../src/PluralResolver';
 import Interpolator from '../../src/Interpolator';
+import logger from '../../src/logger';
 
 describe('Translator', () => {
   describe('translate() with returnObjects=true', () => {
@@ -93,22 +94,24 @@ describe('Translator', () => {
       });
     });
   });
+
   describe('translate() with returnObjects=false', () => {
     let t;
+    let loggerStub;
     let rs;
     let lu;
+    const loggerSpy = sinon.spy();
     before(() => {
+      loggerStub = sinon.stub(logger, 'create');
+      loggerStub.returns({
+        warn: loggerSpy,
+        log: sinon.spy(),
+      });
+
       rs = new ResourceStore({
         en: {
           common: {
             array: ['common_array_en_1', 'common_array_en_2'],
-            object: {
-              value: 'common_object_en_value',
-            },
-            array_with_context: ['lorem ipsum', 'lorem ipsum', 'hello {{what}}'],
-          },
-          special: {
-            array: ['special_array_en_1', 'special_array_en_2'],
           },
         },
       });
@@ -135,44 +138,42 @@ describe('Translator', () => {
       t.changeLanguage('en');
     });
 
-    var tests = [
-      {
-        args: ['common:array'],
-        expected: ['array', ['common_array_en_1', 'common_array_en_2'], { ns: ['common'] }],
-      },
-      {
-        args: ['array'],
-        expected: ['array', ['common_array_en_1', 'common_array_en_2'], { ns: ['common'] }],
-      },
-      {
-        args: ['common:array_with_context', { what: 'world' }],
-        expected: [
-          'array_with_context',
-          ['lorem ipsum', 'lorem ipsum', 'hello {{what}}'],
-          { what: 'world', ns: ['common'] },
-        ],
-      },
-      {
-        args: ['common:object', { what: 'world' }],
-        expected: [
-          'object',
-          { value: 'common_object_en_value' },
-          { what: 'world', ns: ['common'] },
-        ],
-      },
-      {
-        args: ['special:array'],
-        expected: ['array', ['special_array_en_1', 'special_array_en_2'], { ns: ['special'] }],
-      },
-    ];
-
     describe('and "returnedObjectHandler" defined', () => {
-      tests.forEach(test => {
-        it('correctly translates for ' + JSON.stringify(test.args) + ' args', () => {
-          expect(JSON.stringify(t.translate.apply(t, test.args))).to.eql(
-            JSON.stringify(test.expected),
-          );
-        });
+      afterEach(() => {
+        loggerStub.reset();
+      });
+
+      it('should not emit a warning', () => {
+        t.translate.apply(t, ['common:array']);
+        expect(
+          loggerSpy.calledWith('accessing an object - but returnObjects options is not enabled!'),
+        ).to.be.false;
+      });
+    });
+
+    describe('and "returnedObjectHandler" is not defined', () => {
+      it('should emit a warning', () => {
+        t = new Translator(
+          {
+            resourceStore: rs,
+            languageUtils: lu,
+            pluralResolver: new PluralResolver(lu, { prepend: '_', simplifyPluralSuffix: true }),
+            interpolator: new Interpolator(),
+          },
+          {
+            keySeparator: '.',
+            contextSeparator: '_',
+            returnObjects: false,
+            ns: ['common', 'special'],
+            defaultNS: 'common',
+            interpolation: {},
+          },
+        );
+        t.changeLanguage('en');
+        t.translate.apply(t, ['common:array']);
+        expect(
+          loggerSpy.calledWith('accessing an object - but returnObjects options is not enabled!'),
+        ).to.be.true;
       });
     });
   });